Move error handler tests to internal/errorhandler
