+++ internal/errorhandler/errorhandler_test.go
@@ -9,8 +9,30 @@ import (
 	"log"
 	"os"
 	"strings"
+	"sync"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
+
+type fnErrHandler func(error)
+
+func (f fnErrHandler) Handle(err error) { f(err) }
+
+var noopEH = fnErrHandler(func(error) {})
+
+type nonComparableErrorHandler struct {
+	ErrorHandler
+
+	nonComparable func() //nolint:unused  // This is not called.
+}
+
+func resetForTest(t testing.TB) {
+	t.Cleanup(func() {
+		globalErrorHandler = defaultErrorHandler()
+		delegateErrorHandlerOnce = sync.Once{}
+	})
+}
 
 func TestErrDelegator(t *testing.T) {
 	buf := new(bytes.Buffer)
@@ -29,7 +51,7 @@ func TestErrDelegator(t *testing.T) {
 	buf.Reset()
 
 	var gotErr error
-	e.SetDelegate(fnErrHandler(func(e error) { gotErr = e }))
+	e.setDelegate(fnErrHandler(func(e error) { gotErr = e }))
 	e.Handle(err)
 
 	if buf.String() != "" {
@@ -39,6 +61,53 @@ func TestErrDelegator(t *testing.T) {
 	}
 }
 
-type fnErrHandler func(error)
+func TestSetErrorHandler(t *testing.T) {
+	t.Run("Set With default is a noop", func(t *testing.T) {
+		resetForTest(t)
+		SetErrorHandler(GetErrorHandler())
 
-func (f fnErrHandler) Handle(err error) { f(err) }
+		eh, ok := GetErrorHandler().(*ErrDelegator)
+		if !ok {
+			t.Fatal("Global ErrorHandler should be the default ErrorHandler")
+		}
+
+		if eh.delegate.Load() != nil {
+			t.Fatal("ErrorHandler should not delegate when setting itself")
+		}
+	})
+
+	t.Run("First Set() should replace the delegate", func(t *testing.T) {
+		resetForTest(t)
+
+		SetErrorHandler(noopEH)
+
+		_, ok := GetErrorHandler().(*ErrDelegator)
+		if ok {
+			t.Fatal("Global ErrorHandler was not changed")
+		}
+	})
+
+	t.Run("Set() should delegate existing ErrorHandlers", func(t *testing.T) {
+		resetForTest(t)
+
+		eh := GetErrorHandler()
+		SetErrorHandler(noopEH)
+
+		errDel, ok := eh.(*ErrDelegator)
+		if !ok {
+			t.Fatal("Wrong ErrorHandler returned")
+		}
+
+		if errDel.delegate.Load() == nil {
+			t.Fatal("The ErrDelegator should have a delegate")
+		}
+	})
+
+	t.Run("non-comparable types should not panic", func(t *testing.T) {
+		resetForTest(t)
+
+		eh := nonComparableErrorHandler{}
+		assert.NotPanics(t, func() { SetErrorHandler(eh) }, "delegate")
+		assert.NotPanics(t, func() { SetErrorHandler(eh) }, "replacement")
+	})
+}
