+++ internal/global/state_test.go
@@ -15,12 +15,6 @@ import (
 	tracenoop "go.opentelemetry.io/otel/trace/noop"
 )
 
-type nonComparableErrorHandler struct {
-	ErrorHandler
-
-	nonComparable func() //nolint:unused  // This is not called.
-}
-
 type nonComparableTracerProvider struct {
 	trace.TracerProvider
 
@@ -31,63 +25,6 @@ type nonComparableMeterProvider struct {
 	metric.MeterProvider
 
 	nonComparable func() //nolint:unused  // This is not called.
-}
-
-type fnErrHandler func(error)
-
-func (f fnErrHandler) Handle(err error) { f(err) }
-
-var noopEH = fnErrHandler(func(error) {})
-
-func TestSetErrorHandler(t *testing.T) {
-	t.Run("Set With default is a noop", func(t *testing.T) {
-		ResetForTest(t)
-		SetErrorHandler(GetErrorHandler())
-
-		eh, ok := GetErrorHandler().(*ErrDelegator)
-		if !ok {
-			t.Fatal("Global ErrorHandler should be the default ErrorHandler")
-		}
-
-		if eh.Delegate.Load() != nil {
-			t.Fatal("ErrorHandler should not delegate when setting itself")
-		}
-	})
-
-	t.Run("First Set() should replace the delegate", func(t *testing.T) {
-		ResetForTest(t)
-
-		SetErrorHandler(noopEH)
-
-		_, ok := GetErrorHandler().(*ErrDelegator)
-		if ok {
-			t.Fatal("Global ErrorHandler was not changed")
-		}
-	})
-
-	t.Run("Set() should delegate existing ErrorHandlers", func(t *testing.T) {
-		ResetForTest(t)
-
-		eh := GetErrorHandler()
-		SetErrorHandler(noopEH)
-
-		errDel, ok := eh.(*ErrDelegator)
-		if !ok {
-			t.Fatal("Wrong ErrorHandler returned")
-		}
-
-		if errDel.Delegate.Load() == nil {
-			t.Fatal("The ErrDelegator should have a delegate")
-		}
-	})
-
-	t.Run("non-comparable types should not panic", func(t *testing.T) {
-		ResetForTest(t)
-
-		eh := nonComparableErrorHandler{}
-		assert.NotPanics(t, func() { SetErrorHandler(eh) }, "delegate")
-		assert.NotPanics(t, func() { SetErrorHandler(eh) }, "replacement")
-	})
 }
 
 func TestSetTracerProvider(t *testing.T) {
