+++ internal/global/util_test.go
@@ -6,14 +6,11 @@ package global
 import (
 	"sync"
 	"testing"
-
-	"go.opentelemetry.io/otel/internal/errorhandler"
 )
 
 // ResetForTest configures the test to restores the initial global state during
 // its Cleanup step.
 func ResetForTest(t testing.TB) {
-	errorhandler.ResetForTest(t)
 	t.Cleanup(func() {
 		globalTracer = defaultTracerValue()
 		globalPropagators = defaultPropagatorsValue()
